EMPCA functions

ExpectationMaximizationPCA.EMPCA! — Method

EMPCA!(lm, data_tmp, weights; kwargs...)

Perform Expectation maximization PCA on data_temp. See https://github.com/christiangil/ExpectationMaximizationPCA.jl

StellarSpectraObservationFitting.DEMPCA! — Method

DEMPCA!(M, scores, rv_scores, μ, data_temp, weights, doppler_comp; min_flux=0., max_flux=2., save_doppler_in_M1=true, kwargs...)

Perform Doppler-constrained Expectation maximization PCA on data_temp

StellarSpectraObservationFitting.doppler_component — Method

doppler_component(λ, flux)

Estimate the basis vector that encodes the effects of a doppler shift based on Taylor expanding f(λ/(1 + z)) about z=0 doppler_comp = λ * dF/dλ -> units of flux

StellarSpectraObservationFitting.doppler_component_AD — Method

doppler_component_AD(λ, flux)

Estimate the basis vector that encodes the effects of a doppler shift based on Taylor expanding f(λ/(1 + z)) about z=0. Autodiff friendly doppler_comp = λ * dF/dλ -> units of flux

StellarSpectraObservationFitting.project_doppler_comp! — Method

project_doppler_comp!(scores, data_temp, doppler_comp, weights)

Finding the optimal scores to remove the weighted projection of doppler_comp from data_temp

StellarSpectraObservationFitting.project_doppler_comp — Method

project_doppler_comp(data_temp, doppler_comp, weights)

Remove the weighted projection of doppler_comp from data_temp

StellarSpectraObservationFitting.simple_derivative — Method

simple_derivative(x)

Estimate the derivative of x with finite differences (assuming unit separation)

StellarSpectraObservationFitting.simple_derivative_AD — Method

simple_derivative_AD(x)

Estimate the derivative of x with finite differences (assuming unit separation). Autodiff friendly