Index

StellarSpectraObservationFitting.doppler_component

StellarSpectraObservationFitting.find_modes

StellarSpectraObservationFitting.general_lst_sq

StellarSpectraObservationFitting.multiple_append!

StellarSpectraObservationFitting.ordinary_lst_sq

StellarSpectraObservationFitting.simple_derivative

StellarSpectraObservationFitting.trapz_small

StellarSpectraObservationFitting.doppler_component — Method

Estimate the basis vector that encodes the effects of a doppler shift based on Taylor expanding f(λ/(1 + z)) about z=0 doppler_comp = λ * dF/dλ -> units of flux

StellarSpectraObservationFitting.find_modes — Method

Return an amount of indices of local maxima of a data array

StellarSpectraObservationFitting.general_lst_sq — Method

Solve a linear system of equations (optionally with variance values at each point or covariance array) see (https://en.wikipedia.org/wiki/Generalizedleastsquares#Method_outline)

StellarSpectraObservationFitting.multiple_append! — Method

a generalized version of the built in append!() function

StellarSpectraObservationFitting.ordinary_lst_sq — Method

Solve a linear system of equations (optionally with variance values at each point or covariance array) see (https://en.wikipedia.org/wiki/Generalizedleastsquares#Method_outline)

StellarSpectraObservationFitting.simple_derivative — Method

Estimate the derivatives of a vector

StellarSpectraObservationFitting.trapz_small — Method

trapezoidal integration, shamelessly modified from. Faster for small ranges https://github.com/dextorious/NumericalIntegration.jl/blob/master/src/NumericalIntegration.jl